+++ src/index.js
@@ -12,8 +12,8 @@ import Navbar from './components/Navbar/Navbar';
 import GoogleMap from './components/GoogleMap/GoogleMap';
 import BookItem from './components/BookItem/BookItem';
 
-import AccountPage from './components/Account/Account';
 import AdminPage from './components/Admin/Admin';
+import AccountPage from './pages/AccountPage';
 import SignInPage from './pages/SignInPage';
 import SignUpPage from './pages/SignUpPage';
 import PasswordForgetPage from './pages/PasswordForgetPage';
